backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain-core
httpx
pydantic
